Receive AI decisions through IEnumerable, not through an IRover interface
This is a significant change to the architecture of AI interaction. Instead of an AI manipulating the rover through an interface of IRover, it returns an IEnumerable<RoverAction>. The AI is expected to use "yield return" to provide an action. The action is then applied to the Rover, after which the next action is retrieved from the AI. The AI can access rover state through IRoverStatusAccessor.

There are several benefits to this method. Primarily, it removes the requirement for the AI to run asynchronously from an interactive application, as each individual step is controlled by the AI. Similarly, this means the application has total control of how the quickly the AI is run - for example, the user can watch each step individually.
Additionally, there is no need to use an OutOfMovesOrPowerException to break out of AIs that don't properly check for halting. Instead, the simulator can check after applying each action and simply stop enumerating actions.

There are a couple downsides as well. The overhead of the IEnumerable state machine seems to slow down simulation to roughly half the previous speed. Also, the ScratchAis are less easily translatable to Scratch, since they also need to use "yield return" instead of interfacing with the rover.

There may still exist a bug in this implementation: the AIs 'Random', 'Intelligent Random', and 'Mark I' all transmit more samples on average than previously, and with a lower variance. This is a statistically significant difference, even though none of their behavior has been intentionally modified. Both the Mark II and Fixed State AIs are consistent with earlier results.