Improve Fixed State AI by about 3.4 samples per simulation
We now can gather samples while gathering power. If the rover occupies a smooth tile with more than 1 adjacent smooth tile, it will return to this tile its following move. Then it moves onto an adjacent smooth tile, samples it, and moves back to the central tile. It does this until the central tile has 1 or fewer adjacent smooth tiles.

This allows the rover to collect a limited number of samples while in the power gathering phase, but still gather the same amount of power.