Further prevent the Pathfinding AI from running out of power
The primary way the pathfinding AI gathers power is by finding two adjacent smooth tiles and driving up the no backtracking stat. However, in rare cases, the AI is not able to find such an area before its initial power runs out. With this change, the rover will now very slowly gather power in place if the next move would drain the rest of the power.